pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "hed-matlab"
version = "1.0.0"
description = "MATLAB tools for HED (Hierarchical Event Descriptors)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "HED Working Group", email = "hed.maintainers@gmail.com"}
]
keywords = ["HED", "hierarchical event descriptors", "MATLAB", "neuroscience"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: MATLAB",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[project.urls]
Homepage = "https://github.com/hed-standard/hed-matlab"
Documentation = "https://www.hedtags.org/hed-matlab"
Repository = "https://github.com/hed-standard/hed-matlab"
"Bug Tracker" = "https://github.com/hed-standard/hed-matlab/issues"

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "codespell>=2.2.0",
    "black>=24.0.0",
    "mdformat>=0.7.0",
    "mdformat-myst>=0.1.5",
    "mdformat-tables>=0.4.0",
]
docs = [
    "Sphinx>=5.0,<9.0",
    "myst-parser>=0.18.1,<5.0",
    "furo>=2024.0.0",
    "sphinxcontrib-matlabdomain==0.22.1",
    "sphinx-autodoc-typehints>=1.20.0,<4.0",
    "pygments>=2.10.0",
    "sphinx-design>=0.5.0",
    "sphinx-copybutton>=0.5.0",
    "linkify-it-py>=2.0.0",
]

[tool.sphinx]
source-dir = "docs/source"
build-dir = "docs/_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.setuptools]
packages = []

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
